requirements.txt
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.0
python-telegram-bot==20.7
pytest==7.4.0
schedule==1.2.1
pyyaml==6.0.1